Open source the SSE package.  (#490)
## Summary
Open source the SSE package. Mainly a copy-paste from the code in Axiom,
but I did tweak the validation error stuff to use the Sentinel error.
Once this is in, I'll delete the SSE code in Axiom and have it use this
code instead.

## How was it tested?
Ran unit tests

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).